+++ src/domain/Food/Food.js
@@ -1,6 +1,8 @@
-import ERROR_MESSAGE_GENERATOR from '../../constants/error.js';
-import ApplicationError from '../../exceptions/ApplicationError.js';
 import { isBlank } from '../../utils/validator/validator.js';
+
+import ERROR_MESSAGE_GENERATOR from '../../constants/error.js';
+
+import ApplicationError from '../../exceptions/ApplicationError.js';
 
 /**
  * @typedef {import("../../types/price").PriceInfo} PriceInfo
