+++ src/domain/OrderTaker/OrderTaker.js
@@ -1,7 +1,9 @@
-import ERROR_MESSAGE_GENERATOR from '../../constants/error.js';
-import ApplicationError from '../../exceptions/ApplicationError.js';
 import { Appetizer, Dessert, Drink, MainCourse } from '../Food/index.js';
 import OrderDetail from '../OrderDetail/OrderDetail.js';
+
+import ERROR_MESSAGE_GENERATOR from '../../constants/error.js';
+
+import ApplicationError from '../../exceptions/ApplicationError.js';
 
 /**
  * @typedef MenuInfo 메뉴에 대한 정보입니다.
