+++ src/index.js
@@ -1,3 +1,3 @@
-import App from "./App.js";
+import App from './App.js';
 
 new App().run();
